requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.23.2
pydantic==2.1.1
SQLAlchemy==2.0.19
psycopg2==2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.1
pytest==7.4.0
asyncpg==0.28.0
pydantic[email]
python-multipart==0.0.6
starlette==0.27.0
alembic==1.11.2
python-dotenv==1.0.0
pytest-asyncio==0.21.1
pytest-order==1.1.0
pip==23.2.1